workflow/service_chaining/email_notification/requirements.txt
flask>=1.0.2
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
